% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rd_file_path.R
\name{get_help_text}
\alias{get_help_text}
\title{Get the help text}
\usage{
get_help_text(func_name, package = NULL)
}
\arguments{
\item{func_name}{a \code{\link{name}} or \code{\link{character}} string specifying the function for which help is sought.}

\item{package}{a name or character vector giving the packages to look into for documentation, or \code{NULL}}
}
\value{
The plantext of the .Rd file
}
\description{
Similar to \code{help()}, but this function will obtain the text of .Rd file
}
\examples{
cat(get_help_text(sapply))          # Without quotes
cat(get_help_text("sapply"))        # Can use as a string
cat(get_help_text(sapply, base))   # Specify package using name
cat(get_help_text("sapply", "base"))  # Use string to specify package
#cat(get_help_text(sapply, stats))   # Specify package using name, but wrong package
#cat(get_help_text("sapply", "stats"))  # Use string to specify package, but wrong package
}
