% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help.R
\name{Help}
\alias{Help}
\title{Help}
\usage{
Help(
  help.txt,
  language,
  prompt = 1,
  add.prompt = NULL,
  roll = 1,
  api.key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{help.txt}{text from \code{get_help_text}}

\item{language}{a character for specify language. This will be inserted in the prompt}

\item{prompt}{prompt for chatGPT. See details below.}

\item{add.prompt}{Additional prompt for chatGPT. Will be added after \code{prompt}}

\item{roll}{Roll of chatGPT. See \code{\link{reset_chat_session}}}

\item{api.key}{OpenAI api key}
}
\description{
Help
}
\details{
For \code{prompt=1}, the prompt is "Translate the following .Rd file text into \emph{target_language}. Ensure no changes are made to any \{ or \} characters

For \code{roll=1}, the system roll is "You are a helpful assistant familiar with R and Latex."
}
\examples{
#Sys.setenv(OPENAI_API_KEY = "Your OpenAI api key")
#txt = get_help_text(filter)
#Help(txt, language = "Chinese")

}
