% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelpGPT.R
\name{HelpGPT}
\alias{HelpGPT}
\title{HelpGPT}
\usage{
HelpGPT(
  func_name,
  package = NULL,
  language = NULL,
  prompt = 1,
  add.prompt = NULL,
  roll = 1,
  api.key = Sys.getenv("OPENAI_API_KEY"),
  model = c("gpt-4o", "gpt-4o-mini", "gpt-3.5-turbo")
)
}
\arguments{
\item{func_name}{a \code{\link{name}} or \code{\link{character}} string specifying the function for which help is sought.}

\item{package}{a name or character vector giving the packages to look into for documentation, or \code{NULL}.}

\item{language}{a character for specify language. This will be inserted in the pre-specified prompt.}

\item{prompt}{Pre-specified prompt for chatGPT. See details below. User can also use their own prompt character.}

\item{add.prompt}{Additional prompt for chatGPT. Will be added after \code{prompt}.}

\item{roll}{Roll of chatGPT. See \code{\link{reset_chat_session}} in package \code{\link{chatgpt}}.}

\item{api.key}{a character, your OpenAI api key.}

\item{model}{ID of the model to use. Suggest to use \code{"gpt-4o"}. See https://platform.openai.com/docs/models.}
}
\description{
Utilized OpenAI’s API to translate .Rd documentation files into any language.
}
\details{
For \code{prompt=1}, the prompt is:

"Translate the following .Rd file text into formal, polished \if{html}{\out{<TargetLanguage>}}, making sure to maintain a professional tone. Do not add any comment. Ensure no changes are made to any \{ or \} characters, including preserving sequences of multiple \} (like \}\}). Do not alter any LaTeX commands or formatting commands (such as \verb{\\item{}{}}, \verb{\\code{}}, \verb{\\emph{}}, etc.). Each \verb{\\item} should retain its original structure and indentation, and nested items should remain within their original scope. Additionally, avoid translating any R code or variable names."

For \code{roll=1}, the system roll is:

"You are a helpful assistant familiar with R and Latex."
}
\examples{

\dontrun{
Sys.setenv(OPENAI_API_KEY = "Your OpenAI api key")
HelpGPT(filter, stats, language = "Spanish")
HelpGPT("sapply", prompt = "Your own prompt")
}

}
